requirements.txt
numpy
pandas
scikit-learn
matplotlib
streamli